requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-community
google-generativeai==0.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
PyPDF2==3.0.1
chromadb==0.4.22
huggingface-hub==0.20.3